refactor : 안쓰는 컴포넌트 정리, 북마크 아이콘 수정(#171)
